==> R/03_State_popup.Rmd <==
---
title: "Pop-ups"
author: "CONAPO"
output:
   html_document:
      highlight: tango
      theme: flatly
      toc: yes
      toc_depth: 2
      toc_float:
        collapsed: yes
---
\usepackage{color}

```{=html}
<style>
code.r{
  font-size: 10px;
}
pre {
  font-size: 12px
}
</style>

<style>
body {
text-align: justify;
font-style: normal;
font-family: "Century Gothic";
font-size: 12px
}
h1.title {
  font-size: 40px;
  color: #000D3B;
}
h1 {
  color: #B6854D;
}
h2 {
  color: #172984;
}
h3 {
  color: #172984;
}
</style>
```

```{=html}
<style>
.nav>li>a {
    position: relative;
    display: block;
    padding: 10px 15px;
    color: #0A2687;
}
.nav-pills>li.active>a, .nav-pills>li.active>a:hover, .nav-pills>li.active>a:focus {
    color: #ffffff;
    background-color: #09C2BC;
}
</style>
```

```{r, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE, cache = FALSE, cache.lazy = FALSE, 
                         eval = FALSE, class.source = "fold-show")
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
options(digits = 2, encoding = "UTF8")
```   
 

```{r, echo=FALSE}
rm(list = ls())
```

```{r, echo=FALSE}
setwd(here::here())
```


```{r}
require(dplyr)
require(forcats)
require(tibble)
require(sf)
require(sp)
require(stringr)
require(dplyr)
require(leaflet)
require(RColorBrewer)
```

# Base de datos   

```{r}
load(paste0(here::here(), "/Output/ICE_2020.RData"))
```


# Shapefiles 

```{r}
## Shapefiles
shape_municipios_1 <- geojsonio::geojson_read(paste0(here::here(), "/Output/Shapes/municipios_1_json.geojson"), what = "sp") 
shape_municipios_2 <- geojsonio::geojson_read(paste0(here::here(), "/Output/Shapes/municipios_2_json.geojson"), what = "sp") 
shape_municipios_3 <- geojsonio::geojson_read(paste0(here::here(), "/Output/Shapes/municipios_3_json.geojson"), what = "sp") 
shape_municipios_4 <- geojsonio::geojson_read(paste0(here::here(), "/Output/Shapes/municipios_4_json.geojson"), what = "sp") 
shape_municipios_5 <- geojsonio::geojson_read(paste0(here::here(), "/Output/Shapes/municipios_5_json.geojson"), what = "sp") 
```


# Paleta de colores 

```{r}
require(RColorBrewer)
paleta <- rev(brewer.pal(5, "RdPu"))
paleta <- c("#7A0177", "#C51B8A", "#D45085", "#E0979C", "#CDAD9E")

mypalette <- leaflet::colorFactor(palette = paleta, 
                                   domain = forcats::fct_relevel(ICE_2020$G_ICE, c("Completa", "Grave", "Moderada", "Ligera", "No hay")), 
                                    na.color = "transparent")
```


## Popup `Completa`

```{r}
## Paleta de colores                         
p <- shape_municipios_1 %>%
      spdplyr:::inner_join.Spatial(., ICE_2020 %>% 
                                        select(CVE_MUN, G_ICE) %>%
                                         filter(G_ICE == "Completa"),
                                          by = c("CVE_MUN")) %>%
      
        leaflet() %>% 
         addTiles() %>% 
          setView(lat = 23.634501, lng = -102.552784, zoom = 5) %>%
           addProviderTiles(providers$Esri.WorldImagery) %>%
            addPolygons(fillColor = ~mypalette(G_ICE))

fillColor <- paleta[1]

df_municipio <- shape_municipios_1@data %>%
                 left_join(., ICE_2020 %>% 
                               mutate(across(.cols = c(4:5, 7, 9, 11), as.numeric)) %>%
                                filter(G_ICE %in% "Completa"),
                                 by = c("CVE_MUN"))
                                                
## Etiquetas del mapa a nivel municipal 
state_popup_municipio <- paste('<font size="3"; font-family: "Century Gothic"><strong>',"Estado:",'</strong>','<strong>', df_municipio$NOM_ENT,'</strong></font>','<br/>',
                               '<font size="2.5"; font-family: "Century Gothic"><strong>',"Municipio:",'</strong>','<strong>', df_municipio$NOM_MUN,'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Población Total:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', formatC(df_municipio$POBTOT, format = "f", big.mark = " ", digits = 0),'</strong></font>','<br/>',
                                 '<font size="2"; font-family: "Century Gothic"><strong>',"Grado de Calidad del Entorno:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', df_municipio$G_ICE,'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Índice de Calidad del Entorno (ICE):",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', format(round(df_municipio$ICE, 2), nsmall = 1),'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Grado de marginación:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', df_municipio$GM_2020,'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Índice de marginación:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', format(round(df_municipio$IM_2020, 2), nsmall = 1),'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Grado de equipamiento:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', df_municipio$G_IE,'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Índice de equipamiento:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', format(round(df_municipio$IE, 2), nsmall = 1),'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Grado de accesibilidad:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', df_municipio$GACC,'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Índice de accesibilidad:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', format(round(df_municipio$IAL, 2), nsmall = 1),'</strong></font>','<br/>'
) %>% 
  lapply(htmltools::HTML)

saveRDS(state_popup_municipio, paste0(here::here(), "/Output/PopUp/state_popup_municipio1.RDS"))
```

## Popup `Grave`

```{r}
## Paleta de colores                         
p <- shape_municipios_2 %>%
      spdplyr:::inner_join.Spatial(., ICE_2020 %>% 
                                       select(CVE_MUN, G_ICE) %>%
                                        filter(G_ICE == "Grave"),
                                         by = c("CVE_MUN")) %>%
      
        leaflet() %>% 
         addTiles() %>% 
          setView(lat = 23.634501, lng = -102.552784, zoom = 5) %>%
           addProviderTiles(providers$Esri.WorldImagery) %>%
            addPolygons(fillColor = ~mypalette(G_ICE))

fillColor <- paleta[2]

df_municipio <- shape_municipios_2@data %>%
                 left_join(., ICE_2020 %>% 
                               mutate(across(.cols = c(4:5, 7, 9, 11), as.numeric)) %>%
                                filter(G_ICE %in% "Grave"),
                                 by = c("CVE_MUN"))

## Etiquetas del mapa a nivel municipal 
state_popup_municipio <- paste('<font size="3"; font-family: "Century Gothic"><strong>',"Estado:",'</strong>','<strong>', df_municipio$NOM_ENT,'</strong></font>','<br/>',
                               '<font size="2.5"; font-family: "Century Gothic"><strong>',"Municipio:",'</strong>','<strong>', df_municipio$NOM_MUN,'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Población Total:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', formatC(df_municipio$POBTOT, format = "f", big.mark = " ", digits = 0),'</strong></font>','<br/>',
                                 '<font size="2"; font-family: "Century Gothic"><strong>',"Grado de Calidad del Entorno:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', df_municipio$G_ICE,'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Índice de Calidad del Entorno (ICE):",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', format(round(df_municipio$ICE, 2), nsmall = 1),'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Grado de marginación:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', df_municipio$GM_2020,'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Índice de marginación:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', format(round(df_municipio$IM_2020, 2), nsmall = 1),'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Grado de equipamiento:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', df_municipio$G_IE,'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Índice de equipamiento:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', format(round(df_municipio$IE, 2), nsmall = 1),'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Grado de accesibilidad:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', df_municipio$GACC,'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Índice de accesibilidad:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', format(round(df_municipio$IAL, 2), nsmall = 1),'</strong></font>','<br/>'
) %>% 
  lapply(htmltools::HTML)

saveRDS(state_popup_municipio,  paste0(here::here(), "/Output/PopUp/state_popup_municipio2.RDS"))
```

## Popup `Moderada`

```{r}
## Paleta de colores                         
p <- shape_municipios_3 %>%
      spdplyr:::inner_join.Spatial(., ICE_2020 %>% 
                                       select(CVE_MUN, G_ICE) %>%
                                        filter(G_ICE == "Moderada"),
                                         by = c("CVE_MUN")) %>%
      
        leaflet() %>% 
         addTiles() %>% 
          setView(lat = 23.634501, lng = -102.552784, zoom = 5) %>%
           addProviderTiles(providers$Esri.WorldImagery) %>%
            addPolygons(fillColor = ~mypalette(G_ICE))

fillColor <- paleta[3]

df_municipio <- shape_municipios_3@data %>%
                 left_join(., ICE_2020 %>% 
                               mutate(across(.cols = c(4:5, 7, 9, 11), as.numeric)) %>%
                                filter(G_ICE %in% "Moderada"),
                                 by = c("CVE_MUN"))

## Etiquetas del mapa a nivel municipal 
state_popup_municipio <- paste('<font size="3"; font-family: "Century Gothic"><strong>',"Estado:",'</strong>','<strong>', df_municipio$NOM_ENT,'</strong></font>','<br/>',
                               '<font size="2.5"; font-family: "Century Gothic"><strong>',"Municipio:",'</strong>','<strong>', df_municipio$NOM_MUN,'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Población Total:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', formatC(df_municipio$POBTOT, format = "f", big.mark = " ", digits = 0),'</strong></font>','<br/>',
                                 '<font size="2"; font-family: "Century Gothic"><strong>',"Grado de Calidad del Entorno:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', df_municipio$G_ICE,'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Índice de Calidad del Entorno (ICE):",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', format(round(df_municipio$ICE, 2), nsmall = 1),'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Grado de marginación:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', df_municipio$GM_2020,'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Índice de marginación:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', format(round(df_municipio$IM_2020, 2), nsmall = 1),'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Grado de equipamiento:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', df_municipio$G_IE,'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Índice de equipamiento:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', format(round(df_municipio$IE, 2), nsmall = 1),'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Grado de accesibilidad:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', df_municipio$GACC,'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Índice de accesibilidad:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', format(round(df_municipio$IAL, 2), nsmall = 1),'</strong></font>','<br/>'
) %>% 
  lapply(htmltools::HTML)

saveRDS(state_popup_municipio,  paste0(here::here(), "/Output/PopUp/state_popup_municipio3.RDS"))
```

## Popup `Ligera`

```{r}
## Paleta de colores                         
p <- shape_municipios_4 %>%
      spdplyr:::inner_join.Spatial(., ICE_2020 %>% 
                                       select(CVE_MUN, G_ICE) %>%
                                        filter(G_ICE == "Ligera"),
                                         by = c("CVE_MUN")) %>%
      
        leaflet() %>% 
         addTiles() %>% 
          setView(lat = 23.634501, lng = -102.552784, zoom = 5) %>%
           addProviderTiles(providers$Esri.WorldImagery) %>%
            addPolygons(fillColor = ~mypalette(G_ICE))

fillColor <- paleta[4]

df_municipio <- shape_municipios_4@data %>%
                 left_join(., ICE_2020 %>% 
                               mutate(across(.cols = c(4:5, 7, 9, 11), as.numeric)) %>%
                                filter(G_ICE %in% "Ligera"),
                                 by = c("CVE_MUN"))

## Etiquetas del mapa a nivel municipal 
state_popup_municipio <- paste('<font size="3"; font-family: "Century Gothic"><strong>',"Estado:",'</strong>','<strong>', df_municipio$NOM_ENT,'</strong></font>','<br/>',
                               '<font size="2.5"; font-family: "Century Gothic"><strong>',"Municipio:",'</strong>','<strong>', df_municipio$NOM_MUN,'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Población Total:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', formatC(df_municipio$POBTOT, format = "f", big.mark = " ", digits = 0),'</strong></font>','<br/>',
                                 '<font size="2"; font-family: "Century Gothic"><strong>',"Grado de Calidad del Entorno:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', df_municipio$G_ICE,'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Índice de Calidad del Entorno (ICE):",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', format(round(df_municipio$ICE, 2), nsmall = 1),'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Grado de marginación:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', df_municipio$GM_2020,'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Índice de marginación:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', format(round(df_municipio$IM_2020, 2), nsmall = 1),'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Grado de equipamiento:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', df_municipio$G_IE,'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Índice de equipamiento:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', format(round(df_municipio$IE, 2), nsmall = 1),'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Grado de accesibilidad:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', df_municipio$GACC,'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Índice de accesibilidad:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', format(round(df_municipio$IAL, 2), nsmall = 1),'</strong></font>','<br/>'
) %>% 
  lapply(htmltools::HTML)

saveRDS(state_popup_municipio,  paste0(here::here(), "/Output/PopUp/state_popup_municipio4.RDS"))
```

## Popup `No hay`

```{r}
## Paleta de colores                         
p <- shape_municipios_5 %>%
      spdplyr:::inner_join.Spatial(., ICE_2020 %>% 
                                       select(CVE_MUN, G_ICE) %>%
                                        filter(G_ICE == "No hay"),
                                         by = c("CVE_MUN")) %>%
      
        leaflet() %>% 
         addTiles() %>% 
          setView(lat = 23.634501, lng = -102.552784, zoom = 5) %>%
           addProviderTiles(providers$Esri.WorldImagery) %>%
            addPolygons(fillColor = ~mypalette(G_ICE))

fillColor <- paleta[5]

df_municipio <- shape_municipios_5@data %>%
                 left_join(., ICE_2020 %>% 
                               mutate(across(.cols = c(4:5, 7, 9, 11), as.numeric)) %>%
                                filter(G_ICE %in% "No hay"),
                                 by = c("CVE_MUN"))

## Etiquetas del mapa a nivel municipal 
state_popup_municipio <- paste('<font size="3"; font-family: "Century Gothic"><strong>',"Estado:",'</strong>','<strong>', df_municipio$NOM_ENT,'</strong></font>','<br/>',
                               '<font size="2.5"; font-family: "Century Gothic"><strong>',"Municipio:",'</strong>','<strong>', df_municipio$NOM_MUN,'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Población Total:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', formatC(df_municipio$POBTOT, format = "f", big.mark = " ", digits = 0),'</strong></font>','<br/>',
                                 '<font size="2"; font-family: "Century Gothic"><strong>',"Grado de Calidad del Entorno:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', df_municipio$G_ICE,'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Índice de Calidad del Entorno (ICE):",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', format(round(df_municipio$ICE, 2), nsmall = 1),'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Grado de marginación:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', df_municipio$GM_2020,'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Índice de marginación:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', format(round(df_municipio$IM_2020, 2), nsmall = 1),'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Grado de equipamiento:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', df_municipio$G_IE,'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Índice de equipamiento:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', format(round(df_municipio$IE, 2), nsmall = 1),'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Grado de accesibilidad:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', df_municipio$GACC,'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Índice de accesibilidad:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', format(round(df_municipio$IAL, 2), nsmall = 1),'</strong></font>','<br/>'
) %>% 
  lapply(htmltools::HTML)

saveRDS(state_popup_municipio,  paste0(here::here(), "/Output/PopUp/state_popup_municipio5.RDS"))
```
