==> R/05_ICE_2020.Rmd <==
---
title: "Índice de calidad del entorno"
output: 
  flexdashboard::flex_dashboard:
    logo: https://raw.githubusercontent.com/IndiceMx/ICE_2020/8af22493c664dd9780c8a926127163a2beb37a33/img/images.svg
    output_dir: docs
    orientation: columns
    vertical_layout: fill
    self_contained: yes
    mode: selfcontained
    social: [ "twitter", "facebook", "linkedin" ]
    source_code:  https://github.com/IndiceMx/ICE_2020
    theme: 
      version: 4
      bg: "#FFFFFF"
      fg: "#FDF7F7" 
      primary: "#132463"
      navbar-bg: "#084869"
      base_font: 
        google: Prompt
      heading_font:
        google: Sen
      code_font:
        google: 
          family: Montserrat
          local: false
---
<style>
img{
     max-width:30px;
     max-height:30px;
     background-color: 'transparent';
}
</style>
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE, cache=TRUE,
                      class.source = "fold-hide")
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
options(encoding = "UTF8")
```


<!--html_preserve-->
<iframe src = "https://indicemx.github.io/ICE_2020/" position= "absolute"; top="0"; left="0"; bottom="0"; right="0"; width="100%"; height="100%"; border="none"; margin="0"; padding="0";  frameborder="0"; overflow="hidden"; display= "block";  min-height= "100vh";>"</iframe>
<!--/html_preserve-->
